requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
ansi2html==1.8.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1694046349000/work
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
bleach==6.0.0
cachetools==5.3.1
catboost==1.2.1.1
category-encoders==2.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==2.2.1
cmaes==0.10.0
cmake==3.25.0
colorlog==6.7.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1691044910542/work
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.2
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.14.5
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fastjsonschema==2.18.0
filelock==3.9.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
ipdb==0.13.13
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1693880262622/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1693579759651/work
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1690896916983/work
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1686775611663/work
jupyterlab-widgets==3.0.9
kaggle==1.5.16
kaleido==0.2.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
lit==15.0.7
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.5
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
ml-dtypes==0.2.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
nbformat==5.9.2
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.0
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
optuna==3.3.0
orjson==3.9.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
patsy==0.5.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.3.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
plotly==5.17.0
plotly-resampler==0.9.1
pmdarima==2.0.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1688565951714/work
protobuf==4.24.4
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycaret==3.1.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1691408637400/work
pyod==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1686601365461/work
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.0
schemdraw==0.15
scikit-base==0.5.1
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post9
sktime==0.21.1
split-folders==0.5.1
SQLAlchemy==2.0.20
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.0
sympy==1.11.1
tbats==1.1.3
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.14.1
tomli==2.0.1
torch==2.0.1+cu118
torch-tb-profiler==0.4.3
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827254365/work
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.34.0
triton==2.0.0
tsdownsample==0.1.2
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wrapt==1.14.1
wurlitzer==3.0.3
xgboost==2.0.0
xxhash==3.3.0
yarl==1.9.2
yellowbrick==1.5
zipp==3.16.2
